Move popup.js and tools.js , Delete components directory
