+++ src/index.js
@@ -1,5 +1,5 @@
 import 'core-js/stable';
 import 'regenerator-runtime/runtime';
 import './style.css';
-import './components/popup.js';
+import './modules/popup.js';
 import './modules/display_list_items';
